Create sorta_working_producer
You get random male or females with a 10 points in a array. I want it to have a property where the 0 points that are established with either X or y or x fill out the respective column of row with the "wave" like 0-9 until the row for x or y is completely filled out and where the X values are the column is fill out from that point. The x or y or X points that are randomly chosen from the 10 spots the parent objects have and retain there position so that a cumulative effect can be realized so that a desirability can be expressed when an agent with a total of 10 of these objects has a chance to pair with any of a #num of possible agents, based on the positions of its 10 spots and the possible partners spots. Knowing that only 5 from each will be passed along and maintain  their relative positions. The desirability being that positions of the 9's cross over where positions of the 0's are in the rows vs the columns. The closer the 9's and 0'd are the more desirable the coupling is for both agents. Where there is same spot in the same rows and or columns this should less desirable or undesired has it limits the variability. Perhaps having calculated combined value of the arrays 1024 possible spots being filled with 5 spots from each parent filling in a row or column these will have cross positions that have a value of 0-9 from the two virtual arrays and these 5 rows and 5 columns at the cross points having a combine sum total of there 0-9 value will be more desired where there is a value of 9. Differences off of 9 either up to 18 or down to 0 will be less and less desirable. I think is what I am trying to do.